+++ db/migrate/20260620230000_add_images_only_to_feeds.rb
@@ -1,0 +1,5 @@
+class AddImagesOnlyToFeeds < ActiveRecord::Migration[8.2]
+  def change
+    add_column :feeds, :images_only, :boolean, default: false, null: false
+  end
+end